Style: 컴포넌트 분류로 인한 레이아웃 수정
