+++ src/components/Header.tsx
@@ -9,7 +9,7 @@ export const Header = () => {
         <div className={styles.logoCover}>
           <Image width={170} height={90} src="/images/logo.svg" alt="logo" priority />
         </div>
-        <h1>HarryPotter Book</h1>
+        <h1>Harry Potter Book</h1>
       </Link>
     </header>
   );
