+++ src/styles/booklist.module.scss
@@ -1,0 +1,83 @@
+@import '@styles/mixin';
+
+.bookWrapper {
+  width: 100%;
+  @include flex-vertical(3rem);
+  align-items: center;
+  justify-content: center;
+
+  .searchWrapper {
+    width: 100%;
+    @include center-xy(1rem);
+    justify-content: center;
+
+    & input {
+      max-width: 50rem;
+      width: 100%;
+    }
+
+    & button {
+      display: block;
+      width: 6rem;
+      height: 4.4rem;
+      background-color: var(--color-primary);
+      border-radius: 1.4rem;
+      @include center-xy(0rem);
+      transition: all 0.2s;
+
+      &:hover {
+        background-color: rgb(186, 154, 106);
+      }
+    }
+  }
+
+  .bookList {
+    @include center-xy(3rem);
+    flex-wrap: wrap;
+
+    .bookItem {
+      @include flex-vertical(1rem);
+      width: calc((1200px - 40px - 4 * 30px) / 5);
+      min-height: 38rem;
+
+      .bookCover {
+        position: relative;
+        @include center-xy;
+        width: 100%;
+        aspect-ratio: 216 / 267;
+        flex-shrink: 0;
+      }
+
+      & h2 {
+        font-size: 1.6rem;
+      }
+
+      & p {
+        & span {
+          font-weight: 700;
+        }
+      }
+    }
+  }
+
+  // Pagination
+  .pagination {
+    margin-top: 1.5rem;
+    display: flex;
+    gap: 2rem;
+    justify-content: center;
+    color: #fff;
+
+    & li {
+      & span {
+        color: #888;
+      }
+    }
+
+    .selected {
+      box-shadow: 0 0.2rem 0 0 var(--color-primary);
+      color: var(--color-primary);
+      font-weight: var(--font-semibold);
+    }
+  }
+}
